+++ src/Services/AbstractServiceBuilder.php
@@ -10,18 +10,12 @@ use Psr\Log\LoggerInterface;
 
 abstract class AbstractServiceBuilder
 {
-    protected CoreInterface $core;
-    protected LoggerInterface $log;
     /**
      * @var array<string, mixed>
      */
     protected array $serviceCache;
 
-    public function __construct(
-        CoreInterface $core,
-        LoggerInterface $log
-    ) {
-        $this->core = $core;
-        $this->log = $log;
+    public function __construct(protected CoreInterface $core, protected LoggerInterface $log)
+    {
     }
 }